requirements.txt
absl-py==1.2.0
appdirs==1.4.4
black==22.8.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
docker-pycreds==0.4.0
docstring-parser==0.15
fonttools==4.39.3
fsspec==2022.10.0
fvcore==0.1.5.post20220512
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.47.0
h5py==3.7.0
idna==3.4
iopath==0.1.10
isort==5.10.1
joblib==1.1.0
jsonargparse>=4.17.0
kiwisolver==1.4.4
lightning-bolts==0.7.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.1
mypy-extensions==0.4.3
numpy~=1.23
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
pathtools==0.1.2
Pillow==9.3.0
pip-autoremove==0.10.0
pipdeptree==2.7.0
platformdirs==2.5.2
portalocker==2.5.1
protobuf==3.19.4
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.9.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
sentry-sdk==1.9.3
setproctitle==1.3.1
six==1.16.0
smmap==5.0.0
tabulate==0.8.10
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
timm==0.6.7
tomli==2.0.1
torchmetrics==0.9.3
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.12
wandb==0.16.2
Werkzeug==2.2.3
yacs==0.1.8